Check the opcode of the frame
We implement a system of opcode checking so we can differentiate the
types of frame we are receiving.

In case the opcode is unknown, we close the connection according to RFC
6455: "If an unknown opcode is received, the receiving endpoint MUST
_Fail the WebSocket Connection_."

However, if the opcode is know but we do not support it, we just display
 a warning and ignore the frame.

For now on, nothing happens if we receive a PING or PONG frame.

fixed #6

Signed-off-by: Olivier Gayot <duskcoder@gmail.com>