Handle the PING frames
Whenever we receive a PING frame, we have to reply with a PONG frame
with a payload identical to the one present in the received PING.

we reuse the function send_text() to send the pong frame but add the
opcode to be sent as an argument. This opcode defaults to OPCODE_TEXT so
we do not break the compatibility.

fixes #5

Signed-off-by: Olivier Gayot <duskcoder@gmail.com>