Fixes accessing clients across websocket instances
This corrects the issue with the send_message_to_all method sending messages to clients that are not from the websocket instance the method was called from.

I am not sure as to why it was designed this way. But to keep the ability to send the messages I made the send_message_to_all a class method that gets overridden by a class instance (only for that instance). 

if you wanted to send a message to all clients on all instances you would use WebsocketServer.send_message_to_all. and instance.send_message_to_all will send to only the clients that are connected to the instance that was used in the call. 

I also mangled the class level devices container and added a devices property. this way if the devices container is accessed from an instance only the clients specific to that instance are returned.